moved the config, and added documentation about the config
