+++ src/.vuepress/styles/index.scss
@@ -14,3 +14,8 @@
   --vp-c-accent: #b469ff;
   --vp-c-accent-bg: #b469ff;
 }
+
+
+.token.comment {
+    color: rgb(166, 136, 190) !important
+}
