+++ src/lights/dmx/kistan_fixtures.md
@@ -1,4 +1,4 @@
-# Fixtures in Kistan
+# Fixtures in Kistan's LMixer
 ## Fixtures
 <FixtureInfo relativePathToFixtureType="kistan_fixture_types.html"></FixtureInfo>
